Timer custom implementation

+++ app/src/main/java/io/github/aleksandersh/simpletimer/data/TimerRepository.kt
@@ -4,7 +4,7 @@ import io.reactivex.Observable
 
 interface TimerRepository {
 
-    fun setTime(time: Long)
+    fun setTime(time: Int)
 
-    fun observeTime(): Observable<Long>
+    fun observeTime(): Observable<Int>
 }
